fix number of cpu_msg_handlers calculation

+++ src/core/cpu.c
@@ -50,7 +50,7 @@ void cpu_init(cpuid_t cpu_id)
         cpu_sync_init(&cpu_glb_sync, platform.cpu_num);
 
         ipi_cpumsg_handler_num =
-            (size_t)(_ipi_cpumsg_handlers_id_end - _ipi_cpumsg_handlers_id_start) / sizeof(size_t);
+            ((size_t)_ipi_cpumsg_handlers_id_end - (size_t)_ipi_cpumsg_handlers_id_start) / sizeof(size_t);
         for (size_t i = 0; i < ipi_cpumsg_handler_num; i++) {
             ((size_t*)_ipi_cpumsg_handlers_id_start)[i] = i;
         }
